# web6/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FundedFrens API — Solana meme coin prop trading platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users with optional search
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated user list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserList"
  /v1/admin/users/{userId}:
    get:
      operationId: getAdminUser
      tags: [admin]
      summary: Get a specific user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Admin user detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /v1/admin/orders:
    get:
      operationId: listAdminOrders
      tags: [admin]
      summary: List all orders (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Paginated order list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderList"
...
  /v1/admin/challenge-plans:
    get:
      operationId: listAdminChallengePlans
      tags: [admin]
      summary: List all challenge plans including inactive (admin)
      responses:
        "200":
          description: All challenge plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChallengePlan"
    post:
      operationId: createAdminChallengePlan
      tags: [admin]
      summary: Create a new challenge plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChallengePlanInput"
      responses:
        "200":
          description: Created challenge plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengePlan"
  /v1/admin/challenge-plans/{planId}:
    patch:
      operationId: updateAdminChallengePlan
      tags: [admin]
      summary: Update a challenge plan
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChallengePlanUpdate"
      responses:
        "200":
          description: Updated challenge plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengePlan"
  /v1/admin/payouts:
    get:
      operationId: listAdminPayouts
      tags: [admin]
      summary: List all payout requests (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: All payout requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payout"
  /v1/admin/payouts/{payoutId}/approve:
    post:
      operationId: approveAdminPayout
      tags: [admin]
      summary: Approve a payout request
      parameters:
        - name: payoutId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayoutApproveInput"
      responses:
        "200":
          description: Approved payout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payout"
  /v1/admin/payouts/{payoutId}/reject:
    post:
      operationId: rejectAdminPayout
      tags: [admin]
      summary: Reject a payout request
      parameters:
        - name: payoutId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayoutRejectInput"
      responses:
        "200":
          description: Rejected payout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payout"
  /v1/admin/referrals:
    get:
      operationId: listAdminReferrals
      tags: [admin]
      summary: List all referrals (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Referrals list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralList"
components:
  schemas:
...
    ChallengePlan:
      type: object
      required: [id, name, slug, price_usd, funded_sol, status, display_order, profit_target_pct, max_drawdown_pct, daily_drawdown_pct, min_trading_days, max_position_size_pct, max_open_positions, reactivation_cost_pct]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        price_usd:
          type: number
        funded_sol:
          type: number
        funded_usd_estimate:
          type: ["number", "null"]
        status:
          type: string
          enum: [active, inactive, coming_soon, archived]
        display_order:
          type: integer
        description:
          type: ["string", "null"]
        profit_target_pct:
          type: number
        max_drawdown_pct:
          type: number
        daily_drawdown_pct:
          type: number
        min_trading_days:
          type: integer
        max_position_size_pct:
          type: number
        max_open_positions:
          type: integer
        reactivation_cost_pct:
          type: number
        created_at:
          type: ["string", "null"]
        updated_at:
          type: ["string", "null"]
    ChallengePlanInput:
      type: object
      required: [name, slug, price_usd, funded_sol, profit_target_pct, max_drawdown_pct, daily_drawdown_pct, min_trading_days, max_position_size_pct, max_open_positions, reactivation_cost_pct]
      properties:
        name:
          type: string
        slug:
          type: string
        price_usd:
          type: number
        funded_sol:
          type: number
        status:
          type: string
          enum: [active, inactive, coming_soon, archived]
        display_order:
          type: integer
        description:
          type: string
        profit_target_pct:
          type: number
        max_drawdown_pct:
          type: number
        daily_drawdown_pct:
          type: number
        min_trading_days:
          type: integer
        max_position_size_pct:
          type: number
        max_open_positions:
          type: integer
        reactivation_cost_pct:
          type: number
    ChallengePlanUpdate:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        price_usd:
          type: number
        funded_sol:
          type: number
        status:
          type: string
          enum: [active, inactive, coming_soon, archived]
        display_order:
          type: integer
        description:
          type: string
        profit_target_pct:
          type: number
        max_drawdown_pct:
          type: number
        daily_drawdown_pct:
          type: number
        min_trading_days:
          type: integer
        max_position_size_pct:
          type: number
        max_open_positions:
          type: integer
        reactivation_cost_pct:
          type: number
...
    Order:
      type: object
      required: [id, user_id, status, amount, currency, created_at, updated_at]
      properties:
        id:
          type: integer
        user_id:
          type: string
        challenge_plan_id:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, waiting_for_payment, payment_detected, confirming, confirmed, activated, expired, cancelled, failed]
        amount:
          type: number
        currency:
          type: string
        plan_name:
          type: ["string", "null"]
        plan_slug:
          type: ["string", "null"]
        expected_sol:
          type: ["number", "null"]
        sol_price_usd:
          type: ["number", "null"]
        treasury_wallet:
          type: ["string", "null"]
        expiry_time:
          type: ["string", "null"]
        payment_status:
          type: ["string", "null"]
        tx_signature:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: string
    OrderList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    Payout:
      type: object
      required: [id, user_id, amount, wallet_address, status, created_at]
      properties:
        id:
          type: integer
        user_id:
          type: string
        amount:
          type: number
        wallet_address:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected, paid, cancelled]
        tx_signature:
          type: ["string", "null"]
        rejection_reason:
          type: ["string", "null"]
        created_at:
          type: string
...
    AdminUser:
      type: object
      required: [id, email, role, created_at]
      properties:
        id:
          type: integer
        auth_user_id:
          type: ["string", "null"]
        email:
          type: string
        username:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        telegram_status:
          type: ["string", "null"]
        telegram_username:
          type: ["string", "null"]
        wallet_address:
          type: ["string", "null"]
        referral_code:
          type: ["string", "null"]
        created_at:
          type: string
    AdminUserList:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
        page:
          type: ["integer", "null"]
        limit:
          type: ["integer", "null"]
...
    ReferralList:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            type: object
            additionalProperties: true
        total:
          type: integer
    PayoutApproveInput:
      type: object
      required: [tx_signature]
      properties:
        tx_signature:
          type: string
    PayoutRejectInput:
      type: object
      required: [rejection_reason]
      properties:
        rejection_reason:
          type: string