# web6/lib/api-zod/src/generated/types/challengePlanUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FundedFrens API — Solana meme coin prop trading platform
 * OpenAPI spec version: 0.1.0
 */
import type { ChallengePlanUpdateStatus } from './challengePlanUpdateStatus';

export interface ChallengePlanUpdate {
  name?: string;
  slug?: string;
  price_usd?: number;
  funded_sol?: number;
  status?: ChallengePlanUpdateStatus;
  display_order?: number;
  description?: string;
  profit_target_pct?: number;
  max_drawdown_pct?: number;
  daily_drawdown_pct?: number;
  min_trading_days?: number;
  max_position_size_pct?: number;
  max_open_positions?: number;
  reactivation_cost_pct?: number;
}