# web6/lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FundedFrens API — Solana meme coin prop trading platform
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  user_id: string;
  /** @nullable */
  challenge_plan_id?: number | null;
  status: OrderStatus;
  amount: number;
  currency: string;
  /** @nullable */
  plan_name?: string | null;
  /** @nullable */
  plan_slug?: string | null;
  /** @nullable */
  expected_sol?: number | null;
  /** @nullable */
  sol_price_usd?: number | null;
  /** @nullable */
  treasury_wallet?: string | null;
  /** @nullable */
  expiry_time?: string | null;
  /** @nullable */
  payment_status?: string | null;
  /** @nullable */
  tx_signature?: string | null;
  created_at: string;
  updated_at: string;
}