# web6/lib/api-zod/src/generated/types/payoutRejectInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FundedFrens API — Solana meme coin prop trading platform
 * OpenAPI spec version: 0.1.0
 */

export interface PayoutRejectInput {
  rejection_reason: string;
}